ci: Move utils and tests aout of api folder
