+++ __tests__/post.test.ts
@@ -1,6 +1,6 @@
 import { test, describe, expect } from "@jest/globals";
 
-import postHandler from "../post/[type]";
+import postHandler from "../api/post/[type]";
 import { requestTestToken } from "../__test-utils/req-test-token";
 process.env.NODE_ENV = "test";
 // import { config } from "dotenv";
