+++ api/get/[type].ts
@@ -1,9 +1,9 @@
 import { VercelRequest, VercelResponse } from "@vercel/node";
-import setHeaders from "../_utils/set-headers";
-import { setupResponseData } from "../_utils/setup-response";
-import { supabase } from "../_utils/supabase";
+import setHeaders from "../../_utils/set-headers";
+import { setupResponseData } from "../../_utils/setup-response";
+import { supabase } from "../../_utils/supabase";
 import type { Point } from "geojson";
-import { verifyRequest } from "../_utils/verify";
+import { verifyRequest } from "../../_utils/verify";
 import { PostgrestError } from "@supabase/supabase-js";
 const queryTypes = [
 	"byid",
