+++ api/index.ts
@@ -1,5 +1,5 @@
 import { VercelRequest, VercelResponse } from "@vercel/node";
-import { setupResponseData } from "./_utils/setup-response";
+import { setupResponseData } from "../_utils/setup-response";
 
 export default async function handler(
 	_request: VercelRequest,
